backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Google AI & Cloud
google-generativeai==0.3.2
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-cloud-firestore==2.14.0
firebase-admin==6.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.0

# Data Processing (Python 3.13 compatible)
numpy==2.1.0
pandas==2.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
